src/main/java/com/example/model/Person.java
package com.example.model;

public class Person {
    private String name;
    private int age;
    private Department department;
    private JobTitle jobTitle;

    public Person(String name, int age) {
        this.name = name;
        this.age = age;
    }

    public Person(String name, int age, Department department, JobTitle jobTitle) {
        this.name = name;
        this.age = age;
        this.department = department;
        this.jobTitle = jobTitle;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public int getAge() {
        return age;
    }

    public void setAge(int age) {
        this.age = age;
    }

    public Department getDepartment() {
        return department;
    }

    public void setDepartment(Department department) {
        this.department = department;
    }

    public JobTitle getJobTitle() {
        return jobTitle;
    }

    public void setJobTitle(JobTitle jobTitle) {
        this.jobTitle = jobTitle;
    }


    @Override
    public String toString() {
        return "Person{" +
                "name = '" + name + '\'' +
                ", age = " + age +
                ", department = " + department +
                ", jobTitle = " + jobTitle +
                '}';
    }
}
